Updated bandwidth estimation implementation (#2335)
This is a port of the libwebrtc BWE implementation as of Chromium 128.

Not enabled by default.

+++ jvb/src/main/kotlin/org/jitsi/videobridge/message/BridgeChannelMessage.kt
@@ -483,7 +483,7 @@ class SourceVideoTypeMessage(
  * (that the client can't determine on its own)
  */
 class ConnectionStats(
-    val estimatedDownlinkBandwidth: Double,
+    val estimatedDownlinkBandwidth: Long,
 ) : BridgeChannelMessage() {
     companion object {
         const val TYPE = "ConnectionStats"
